Improve steam32 support. Improve steam parser funcs.
