+++ .golangci.yml
@@ -32,7 +32,7 @@ linters:
     #- exhaustive
     #- exhaustruct
     - exportloopref
-    - forbidigo
+    #- forbidigo
     - forcetypeassert
     #- funlen
     #- gci
